more tests and fixes
